refactor(fe-arch): API 구조 리팩토링 및 타입 분리 완료, Auth Store 오류 점검 및 수정 중
