+++ src/apis/http.ts
@@ -45,3 +45,4 @@ apiClient.interceptors.request.use(async (config) => {
 
 
 export default apiClient;
+
