--- src/types/auth.ts
@@ -1,19 +1,0 @@
-export interface User {
-  id: number;
-  email?: string; // Make email optional
-  nickname: string;
-  profileImageUrl: string | null;
-  bio?: string;
-  friendsCount: number;
-}
-
-export interface LoginRequestDto {
-  email: string;
-  password: string;
-}
-
-export interface SignupRequestDto {
-  email: string;
-  password: string;
-  nickname: string;
-}
