+++ src/types/auth/user.model.ts
@@ -1,0 +1,9 @@
+export interface User {
+    id: number;
+    email?: string;
+    nickname: string;
+    profileImageUrl: string | null;
+    bio?: string;
+    friendsCount: number;
+}
+
