+++ src/types/common/index.ts
@@ -1,0 +1,8 @@
+/**
+ * 여행 상태 Enum (모델, DTO 모두에서 사용 가능)
+ */
+export enum TripStatus {
+    DRAFT = 'DRAFT',
+    PLANNED = 'PLANNED',
+    COMPLETED = 'COMPLETED',
+  }
